backend/requirements.txt
# Core data science stack
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.13.0
pyarrow==14.0.2
matplotlib==3.10.7
streamlit==1.40.2
plotly==5.24.1

# Firebase + utilities
firebase-admin==6.2.0
python-dotenv==1.0.1
pyyaml==6.0.1
joblib==1.3.2

# Optional dev tooling
ipykernel==6.28.0
jupyterlab==4.0.9
rich==13.6.0
typer==0.9.0

# API server
flask==3.0.0
flask-cors==4.0.0




